requirements.txt
nfcpy
pyspotify
pyyaml
pyalsaaudio
